requirements.txt
langchain==0.0.343
keybert==0.8.3
flair==0.13.0
charset_normalizer==3.3.2
unstructured==0.11.2
unstructured_pytesseract==0.3.12
unstructured_inference==0.7.17
python-pptx==0.6.23
pdf2image==1.16.3
pikepdf==8.7.1
pypdf==3.17.1
pdfminer.six==20221105
opencv-python==4.8.1.78
pandas==2.1.3
python-docx==1.1.0
penpyxl==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
gensim==4.3.2
chromadb==0.4.1
openai==1.3.7
tiktoken==0.5.2
atlassian-python-api==3.41.4
markdownify==0.11.6
python-dotenv==1.0.0
# here we have dependencies for UI
streamlit=1.29.0 
streamlit-modal==0.1.0